refacto python code to keep minimum shared
