+++ packages/backend/src/handlers/events/ingest_pulse/data/intensity_levels.json
