+++ packages/backend/src/handlers/events/ingest_pulse/data/sentiment_adjectives.json
