+++ packages/backend/src/handlers/events/ingest_pulse/models.py
@@ -1,0 +1,4 @@
+class PulseDDBIngestionError(Exception):
+    """Custom exception for pulse DDB Ingestion errors"""
+
+    pass
