+++ packages/backend/src/shared/lambda_layer/python/shared/models/pulse.py
@@ -10,20 +10,6 @@ class PulseCreationError(Exception):
     pass
 
 
-class PulseCreationErrorAlreadyPresent(PulseCreationError):
-    """Exception raised when a pulse with the same ID already exists"""
-
-    def __init__(self, user_id: str):
-        super().__init__(f"Pulse already exists for use {user_id}.")
-        self.user_id = user_id
-
-
-class PulseDDBIngestionError(Exception):
-    """Custom exception for pulse DDB Ingestion errors"""
-
-    pass
-
-
 class PulseBase(BaseModel):
     user_id: str
     intent: str
