+++ packages/backend/src/shared/lambda_layer/python/shared/services/aws.py
